Get rid of theme loader
