--- components/basecomponents/MemoryGame/MemoryGameLoading.module.css
@@ -1,4 +1,0 @@
-.loadingContainer {
-  position: absolute;
-  top: 45vh;
-}
